core/backend/requirements.txt
aio-pika==8.2.5
aiohttp==3.8.3
aiormq==6.4.2
aiosignal==1.3.1
amqp==5.1.1
anyio==3.6.2
asgiref==3.5.2
astroid==2.12.13
async-timeout==4.0.2
attrs==22.2.0
autopep8==1.6.0
billiard==4.1.0
celery==5.3.4
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.3
defusedxml==0.7.1
dill==0.3.6
Django==4.1.4
django-allauth==0.54.0
django-cors-headers==3.13.0
django-filter==23.2
django-invitations==2.0.0
django-rest-auth-bearer==0.1.1
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-stubs==1.13.1
django-stubs-ext==0.7.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docstring-to-markdown==0.11
drf-yasg==1.21.7
elastic-transport==8.4.0
elasticsearch==8.5.3
fastapi==0.88.0
flake8==6.0.0
frozenlist==1.3.3
gunicorn==20.1.0
h11==0.14.0
idna==3.4
inflection==0.5.1
isort==5.11.4
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
kombu==5.3.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.3
mccabe==0.7.0
mock==5.0.1
multidict==6.0.3
mypy==0.991
mypy-extensions==0.4.3
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==22.0
pamqp==3.2.1
parso==0.8.3
pick==2.2.0
platformdirs==2.6.2
pluggy==1.0.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.2
pydocstyle==6.2.2
pyflakes==3.0.1
PyJWT==2.8.0
pylint==2.15.9
python-dateutil==2.8.2
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.7.0
python3-openid==3.2.0
pytoolconfig==1.2.4
pytz==2022.7
PyYAML==6.0.1
redis==4.4.1
requests==2.28.1
requests-oauthlib==1.3.1
rope==1.6.0
sh==1.14.3
simplejson==3.19.1
simplejwt==2.0.1
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
sqlparse==0.4.3
starlette==0.22.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
types-pytz==2022.7.0.0
types-PyYAML==6.0.12.2
typing==3.7.4.3
typing_extensions==4.4.0
tzdata==2023.3
ujson==5.6.0
uritemplate==4.1.1
urllib3==1.26.13
uvicorn==0.20.0
vine==5.0.0
wcwidth==0.2.6
whatthepatch==1.0.3
wrapt==1.14.1
yapf==0.32.0
yarl==1.8.2
